requirements.txt
xarray
netcdf4
numpy
pillow
pyproj
pyresample
bokeh
boto3
